提交职位钱包账户冻结与解冻相关逻辑接口

+++ src/main/java/com/korant/youya/workplace/pojo/dto/job/JobCreateDto.java
@@ -5,6 +5,7 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -104,22 +105,22 @@ public class JobCreateDto {
     /**
      * 推荐奖励
      */
-    private Integer award;
+    private BigDecimal award;
 
     /**
      * 面试奖励分配比例
      */
-    private Integer interviewRewardRate;
+    private BigDecimal interviewRewardRate;
 
     /**
      * 入职奖励分配比例
      */
-    private Integer onboardRewardRate;
+    private BigDecimal onboardRewardRate;
 
     /**
      * 转正奖励分配比例
      */
-    private Integer fullMemberRewardRate;
+    private BigDecimal fullMemberRewardRate;
 
     /**
      * 福利标签集合
